Review octree storage, add Octree class
Octree is now stored in linear memory that can be easily uploaded to GPU for surface construction (Cube Marching on GPU)

+++ src/Game.cpp
@@ -74,7 +74,7 @@ void Game::render()
 	{
 		m_fps = m_fps_counter;
 		m_fps_counter = 0;
-		m_last_fps_time = glfwGetTime();
+		m_last_fps_time = (float) glfwGetTime();
 	}
 
 	// dt
